txgen-mvar: pass NixServiceOptions to keepalive
This took a fair amount of rearrangement to broaden the constant
environment in order to pass the keepalive interval in the
NixServiceOptions around. So a few different things happened:

I. create EnvConsts structure encompassing
  A. AsyncBenchmarkControl TVar (potentially changing to IORef)
  B. IOManager
  C. Maybe NixServiceOptions
  This moves the mutable reference in A. to the Reader environment from
  the State of the ExceptionT Env.Error (RWST EnvConsts () Env IO)
  ActionM monad. The reference stays constant though the referenced
  data changes.
II. pass EnvConsts to runScript and runSelftest
III. update Env.hs and NixService.hs accessors

Some of it represents changing a little of the design of the Env and
ActionM once again even after the prior commits, so a fair amount of
squashing commits that entirely redo earlier commits' changes and
rewriting commit messages will need to be done in the sequel.